refactor: RewardInspector component form to a new component
